requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
icrawler==0.6.6
idna==3.4
lxml==4.9.1
Pillow==9.3.0
pycodestyle==2.9.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.12
